NFSUtil: add claimfh option to find_open()
Also, make all options to find_open() named parameters and add
default values for ipaddr and port. Now the open can be searched
for by either the filename or using claimfh.

Change all calls to find_open() to use named parameters instead
of a mixture of positional and named parameters.