PKTT: add _attrlist to IPv4 object
Also, set the _attrlist as a class attribute
so it will be executed once instead of every
time an object is instantiated.
Add constructor to Flags object.
Change order of variable assignment to reflect
packet structure, including documentation
Remove TOS (Type of Service) since this is
deprecated by DSCP and ECN.