REXEC: add remote procedure module
Provides a set of tools for executing a wide range commands, statements,
expressions or functions on a remote host by running a server process
on the remote host serving requests without disconnecting. This allows
for a sequence of operations to be done remotely and not losing state.
A file could be opened remotely, do some other things and then write
to the same opened file without opening the file again.