LOCK: initial commit for locking tests
Basic locking tests verify that a lock is granted using
various arguments to fcntl. These include blocking and
non-blocking locks, read or write locks, where the file
is opened either for reading, writing or both. It also
checks different ranges including limit conditions.

Non-overlapping tests verity that locks are granted on
both the client under test and a second process or a
remote client when locking the same file.

Overlapping tests verity that a lock is granted on the
client under test and a second process or a remote client
trying to lock the same file will be denied if a
non-blocking lock is issue or will be blocked if a
blocking lock is issue on the second process or remote
client.