IO: add NFS I/O tool
This I/O tool is used to create and manipulate files
of different types. The arguments allow running for a
specified period of time as well as running multiple
processes. Each process modifies a single file at a
time and the file name space is different for each
process so there are no collisions between two
different processes modifying the same file.